feat: adjust navbar and update home page
