+++ src/app/globals.css
@@ -25,9 +25,7 @@
 body {
   font-family: 'Rubik', sans-serif;
   color: rgb(var(--foreground-rgb));
-  background: linear-gradient(to bottom,
-      transparent,
-      rgb(var(--background-end-rgb))) rgb(var(--background-start-rgb));
+  
 }
 
 @layer utilities {
