+++ src/components/EditModal.tsx
@@ -12,7 +12,7 @@ export default function EditModal({
   return (
     <div className="fixed inset-0 bg-black bg-opacity-50 z-50 flex justify-center items-center">
       <div className="bg-white p-4 rounded-lg shadow-lg max-w-sm w-full">
-        <button className="float-right text-lg font-semibold" onClick={onClose}>
+        <button className="float-right text-lg text-black" onClick={onClose}>
           &times;
         </button>
         {children}
